pyproject.toml
[tool.poetry]
name = "raid"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
ansicolors = "^1.1"
progress = "^1.5"
colours = "^0.3.0"
scapy = "^2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
